Add files via upload
This game was created in C language. It's the well known game "Battleship" with only one difference: when all your ships, apart from one, are destroyed, you can move your ship horizontal or vertical (depends on the position we have placed them at start), one position per turn. The ZIP file contains all the files and includes a smartplayer (player_9708_9626), a manual player and a random player. The players that play each time can be changed in the file "game.c" and the number of games to be played can be changed in the file "structures.h" in the line "GAMES_TO_RUN".