Add files via upload
This game is a combination of Hunger Games and Fortnite. Two players, one board, it is getting smaller and smaller every 3 rounds. On the board there are food-elements (gain points), traps (lose points) and weapons (bow, sword, pistol). The players move around the board. If player 1 takes the pistol and is near player 2 (distance smaller than 2 blocks), player 1 can kill player 2. If the board gets smaller than the board where the weapons are, the game is over. If the one player's score is negative, the game is also over. 
Graphics are also used in the game and particularly JavaFx.